feat(services): refactor Gemini services
Move Gemini services out of "./services" and into "./gemini"

This change refactors the Gemini services by moving them out of the "./services" directory and into a new "./gemini" directory. This change is made to improve the organization and maintainability of the codebase.
